backend/smartrecruit_backend/requirements.txt
Django==5.0.4
psycopg2-binary==2.9.9
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
channels==4.0.0
daphne==4.0.0
spacy==3.7.4  # Updated to a newer version with better Python 3.11 support
pandas>=1.4.0  # Ensure support for Python 3.11
nltk>=3.7  # Ensure support for Python 3.11
jsonlines==3.0.0
PyPDF2==3.0.0
